HBX-1661: Move class 'org.hibernate.tool.hbm2x.pojo.IteratorTransformer' to 'org.hibernate.tool.internal.util.IteratorTransformer'
